+++ tests/functional/pytorch/jit/003_jit_sd15.py
@@ -61,7 +61,7 @@ def test_jit_sd15():
     assert re.match(r".*CLIPTextModel.*state=tuned.*TensorRTBackend", history[1])
     assert re.match(r".*UNet2DConditionModel.*state=tuned.*TensorRTBackend", history[2])
     assert re.match(r".*Conv2d.*", history[3])
-    assert re.match(r".*Decoder.*state=tuned.*TorchInductorJitBackend", history[4])
+    assert re.match(r".*Decoder.*state=tuned.*", history[4])
     assert re.match(r".*StableDiffusionSafetyChecker.*state=eager.*", history[5])
 
     logger.info("Testing inference with batch_size=1")